Manage vagrant user dotfiles by importing a profile repo
Instead of dumping files from this repo, I want to seperate concerns
and just use the profile repo that I already maintain.  To do this,
however, I have to figure out how to clone a repository from github
during the vagrant provisioning process. This is the first step in that
direction, but sure the fuck not the last.